+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ## [Unreleased]
 
+### Changed
+
+- Updated to Kotlin 2.3.21
+- Updated to Gradle 9.5.0
+- Updated to Android Gradle 9.2.0
+- Updated to Karma Conventions 1.16.1
+- Updated to Filament 2.1.1
+- Updated to Dokka 2.2.0
+- Updated to Kotlin Wrappers 2026.4.15
+- Updated to OSHI 7.0.1
+
 ### Added
 
 - `AsyncRawSource.asInputStream` for JVM & Android